Move dynamic CSS logic to master layout
