--- branding/server/customBrandingRoute.js
@@ -1,41 +1,0 @@
-import { Router } from 'meteor/iron:router';
-import { Branding } from '/branding/collection';
-
-Router.route('customStyle', {
-  where: 'server',
-  path: '/custom.style.css',
-  action () {
-    const branding = Branding.findOne();
-    let styleString;
-    let primaryColor;
-    let primaryColorText;
-
-
-    if (branding && branding.colors && branding.colors.primary) {
-      primaryColor = branding.colors.primary;
-    }
-
-    if (branding && branding.colors && branding.colors.primaryText) {
-      primaryColorText = branding.colors.primaryText;
-    }
-
-    styleString = `
-.coverphoto {
-  background-color: ${primaryColor} !important;
-}
-
-.main-navigation {
-  background-color: ${primaryColor} !important;
-  color: ${primaryColorText} !important;
-}
-
-.search-box {
-  background-color: ${primaryColor} !important;
-}
-`
-
-    const headers = { 'Content-type': 'text/css' };
-    this.response.writeHead(200, headers);
-    this.response.end(styleString);
-  },
-});
