--- core/client/head.html
@@ -1,3 +1,0 @@
-<head>
-  <link rel="stylesheet" type="text/css" href="/custom.style.css">
-</head>
