+++ core/client/layouts/master_layout/master_layout.js
@@ -18,4 +18,28 @@ Template.masterLayout.helpers({
       return branding;
     }
   },
+  primaryColor () {
+    // Get branding
+    const branding = Branding.findOne();
+
+    let primaryColor;
+
+    if (branding && branding.colors && branding.colors.primary) {
+      primaryColor = branding.colors.primary;
+    }
+
+    return primaryColor;
+  },
+  primaryColorText () {
+    // Get branding
+    const branding = Branding.findOne();
+
+    let primaryColorText;
+
+    if (branding && branding.colors && branding.colors.primaryText) {
+      primaryColorText = branding.colors.primaryText;
+    }
+
+    return primaryColorText;
+  },
 });
